Problem 11.4: Compute the k closest stars

+++ ElementsOfProgrammingInterviews/ElementsOfProgrammingInterviews/Structures/HeapNode.cs
@@ -2,10 +2,10 @@ namespace ElementsOfProgrammingInterviews.Structures
 {
     public class HeapNode<T>
     {
-        public int Priority;
+        public double Priority;
         public T Data { get; }
         
-        public HeapNode(int priority, T data)
+        public HeapNode(double priority, T data)
         {
             Priority = priority;
             Data = data;
